// frontend/src/api/types/topic.ts
/**
 * Generated by orval v8.0.0-rc.2 🍺
 * Do not edit manually.
 * Education platform
 * OpenAPI spec version: 0.0.1
 */
import type { Lesson } from './lesson';
import type { TopicParentId } from './topicParentId';

export interface Topic {
  /** Уникальный идентификатор темы */
  id: string;
  /** Название темы */
  title: string;
  /** ID предмета, к которому относится тема */
  subjectId: string;
  /** Список уроков в теме */
  lessons: Lesson[];
  /** ID родительской темы (для иерархии) */
  parentId?: TopicParentId;
  /** Дата создания записи */
  createdAt: string;
  /** Дата последнего обновления записи */
  updatedAt: string;
}